requirements.txt
Flask==2.2.2
hydra-core==1.3.1
librosa==0.7.2
matplotlib==3.5.3
numpy==1.21.6
omegaconf==2.3.0
optuna==3.0.5
python_Levenshtein==0.20.9
pytorch_lightning==1.1.0
requests==2.28.1
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
sox==1.4.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.0
tqdm==4.64.1
wget==3.2
